Move ranking-constants.def to configdefinitions
