+++ configdefinitions/src/vespa/CMakeLists.txt
@@ -44,6 +44,8 @@ vespa_generate_config(configdefinitions persistence.def)
 install_config_definition(persistence.def vespa.config.content.persistence.def)
 vespa_generate_config(configdefinitions rank-profiles.def)
 install_config_definition(rank-profiles.def vespa.config.search.rank-profiles.def)
+vespa_generate_config(configdefinitions ranking-constants.def)
+install_config_definition(ranking-constants.def vespa.config.search.core.ranking-constants.def)
 vespa_generate_config(configdefinitions routing.def)
 install_config_definition(routing.def cloud.config.routing.def)
 vespa_generate_config(configdefinitions routing-provider.def)
