--- searchcore/src/main/java/com/yahoo/vespa/config/search/package-info.java
@@ -1,7 +1,0 @@
-// Copyright 2018 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-
-@ExportPackage
-package com.yahoo.vespa.config.search;
-
-import com.yahoo.osgi.annotation.ExportPackage;
-
