+++ searchcore/src/tests/proton/proton_config_fetcher/proton_config_fetcher_test.cpp
@@ -5,6 +5,7 @@
 #include <vespa/config-imported-fields.h>
 #include <vespa/config-indexschema.h>
 #include <vespa/config-rank-profiles.h>
+#include <vespa/config-ranking-constants.h>
 #include <vespa/config-summary.h>
 #include <vespa/config-summarymap.h>
 #include <vespa/document/repo/documenttyperepo.h>
@@ -17,7 +18,6 @@
 #include <vespa/searchcore/proton/server/i_proton_configurer.h>
 #include <vespa/searchcore/proton/common/hw_info.h>
 #include <vespa/searchsummary/config/config-juniperrc.h>
-#include <vespa/searchcore/config/config-ranking-constants.h>
 #include <vespa/vespalib/testkit/testapp.h>
 #include <vespa/vespalib/util/varholder.h>
 #include <vespa/config-bucketspaces.h>
