+++ searchcore/src/vespa/searchcore/config/CMakeLists.txt
@@ -9,6 +9,4 @@ vespa_generate_config(searchcore_fconfig fdispatchrc.def)
 install_config_definition(fdispatchrc.def vespa.config.search.core.fdispatchrc.def)
 vespa_generate_config(searchcore_fconfig proton.def)
 install_config_definition(proton.def vespa.config.search.core.proton.def)
-vespa_generate_config(searchcore_fconfig ranking-constants.def)
-install_config_definition(ranking-constants.def vespa.config.search.core.ranking-constants.def)
 vespa_generate_config(searchcore_fconfig hwinfo.def)
